Address more feedback
