+++ samples/src/JustSaying.Sample.Restaurant.KitchenConsole/Program.cs
@@ -1,4 +1,5 @@
 using JustSaying.AwsTools.MessageHandling;
+using JustSaying.Messaging.Compression;
 using JustSaying.Sample.Restaurant.KitchenConsole;
 using JustSaying.Sample.Restaurant.KitchenConsole.Handlers;
 using JustSaying.Sample.Restaurant.Models;
@@ -108,7 +109,7 @@ static async Task Run()
                             .WithWriteConfiguration(w =>
                                 w.CompressionOptions = new PublishCompressionOptions
                                 {
-                                    CompressionEncoding = "gzip,base64", // TODO validate
+                                    CompressionEncoding = ContentEncodings.GzipBase64, // TODO validate
                                     MessageLengthThreshold = 0
                                 });
                     });
