--- src/JustSaying/AwsTools/MessageHandling/Compression/IMessageBodyCompression.cs
@@ -1,8 +1,0 @@
-namespace JustSaying.AwsTools.MessageHandling.Compression;
-
-public interface IMessageBodyCompression
-{
-    string ContentEncoding { get; }
-    string Compress(string messageBody);
-    string Decompress(string messageBody);
-}
