--- src/JustSaying/AwsTools/MessageHandling/Compression/IMessageCompressionRegistry.cs
@@ -1,6 +1,0 @@
-namespace JustSaying.AwsTools.MessageHandling.Compression;
-
-public interface IMessageCompressionRegistry
-{
-    IMessageBodyCompression GetCompression(string contentEncoding);
-}
