--- src/JustSaying/AwsTools/MessageHandling/Compression/MessageCompressionRegistry.cs
@@ -1,16 +1,0 @@
-namespace JustSaying.AwsTools.MessageHandling.Compression;
-
-public sealed class MessageCompressionRegistry : IMessageCompressionRegistry
-{
-    private readonly IList<IMessageBodyCompression> _compressions;
-
-    public MessageCompressionRegistry(IList<IMessageBodyCompression> compressions)
-    {
-        _compressions = compressions;
-    }
-
-    public IMessageBodyCompression GetCompression(string contentEncoding)
-    {
-        return _compressions.FirstOrDefault(x => x.ContentEncoding == contentEncoding);
-    }
-}
