+++ src/JustSaying/AwsTools/MessageHandling/SnsMessagePublisher.cs
@@ -2,8 +2,8 @@ using System.Text;
 using Amazon.Runtime;
 using Amazon.SimpleNotificationService;
 using Amazon.SimpleNotificationService.Model;
-using JustSaying.AwsTools.MessageHandling.Compression;
 using JustSaying.Messaging;
+using JustSaying.Messaging.Compression;
 using JustSaying.Messaging.Interrogation;
 using JustSaying.Messaging.MessageSerialization;
 using JustSaying.Models;
