+++ tests/JustSaying.UnitTests/AwsTools/MessageHandling/Sqs/WhenPublishingAsync.cs
@@ -1,6 +1,5 @@
 using Amazon.SQS.Model;
 using JustSaying.AwsTools.MessageHandling;
-using JustSaying.AwsTools.MessageHandling.Compression;
 using JustSaying.Messaging;
 using JustSaying.Messaging.MessageSerialization;
 using JustSaying.TestingFramework;
